pyproject.toml
[tool.poetry]
name = "garagino"
version = "0.0.1"
description = "Sistema criado para gerenciamento e reserva de salas do laboratorio garagem"
authors = ["Francisco Antonio da Luz Filho <falf@cesar.school>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.6.2"
autopep8 = "1.7.1"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.6"
dnspython = "2.2.1"
email-validator = "1.3.0"
fastapi = "0.85.1"
fastapi-utils = "0.2.1"
greenlet = "1.1.3.post0"
h11 = "0.14.0"
httptools = "0.5.0"
idna = "3.4"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.1"
orjson = "3.8.1"
pycodestyle = "2.9.1"
pydantic = "1.10.2"
pydantic-core = "2.10.1"
python-dotenv = "0.21.0"
python-multipart = "0.0.5"
pyyaml = "6.0"
requests = "2.28.1"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "1.4.42"
starlette = "0.20.4"
tomli = "2.0.1"
typing-extensions = "4.8.0"
ujson = "5.5.0"
urllib3 = "1.26.12"
uvicorn = "0.18.3"
watchfiles = "0.18.0"
websockets = "10.4"
sqlalchemy-utils = "^0.41.1"
pg8000 = "^1.30.3"
python-decouple = "^3.8"
pyjwt = "^2.8.0"
passlib = "^1.7.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
